+++ models/Newsletters.php
@@ -9,6 +9,7 @@ use yii\db\ActiveRecord;
 use yii\behaviors\TimestampBehavior;
 use yii\behaviors\BlameableBehavior;
 use yii\validators\EmailValidator;
+use yii\helpers\FileHelper;
 use wdmg\validators\JsonValidator;
 use wdmg\validators\EmailsValidator;
 use yii\helpers\Json;
@@ -96,45 +97,13 @@ class Newsletters extends ActiveRecord
 
     public function validateRecipients($attribute, $params)
     {
-        $error = false;
         $validator = new EmailsValidator();
         $validator->allowName = true;
 
         $data = \yii\helpers\Json::decode($this->recipients);
         if ($emails = array_values($data)) {
             $validator->validate($emails);
-
-
-            /*foreach ($data as $key => $item) {
-                if (preg_match('/email_id:(\d)/', $key)) {
-                    preg_match("/[\._a-zA-Z0-9-]+@[\._a-zA-Z0-9-]+/i", $item, $matches);
-                    if ($email = $matches[0]) {
-                        if (!$error && !$validator->validate(trim($email))) {
-                            $error = true;
-                        }
-                    }
-                } else if (preg_match('/list_id:(\d)/', $key, $match)) {
-                    if ($list = $this->getSubscribersFromList(['list_id' => intval($match[1])])) {
-                        foreach ($list as $key => $item) {
-                            if (!$error && !$validator->validate(trim($item['email']))) {
-                                $error = true;
-                            }
-                        }
-                    }
-                } else {
-                    preg_match("/[\._a-zA-Z0-9-]+@[\._a-zA-Z0-9-]+/i", $item, $matches);
-                    if ($email = $matches[0]) {
-                        if (!$error && !$validator->validate(trim($email))) {
-                            $error = true;
-                        }
-                    }
-                }
-            }*/
         }
-/*
-        if ($error) {
-            $this->addError($attribute, Yii::t('app/modules/newsletters', 'One or more recipients are not a valid email address.'));
-        }*/
     }
 
     /**
@@ -151,6 +120,18 @@ class Newsletters extends ActiveRecord
         }
         return false;
 
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function afterFind()
+    {
+
+        if (!($this->workflow = Json::decode($this->workflow)))
+            $this->workflow = [];
+
+        parent::afterFind();
     }
 
     /**
@@ -217,7 +198,7 @@ class Newsletters extends ActiveRecord
             $aliases = array_keys($extension['alias']);
             foreach ($aliases as $alias) {
                 if (is_dir(Yii::getAlias($alias . '/mail/layouts'))) {
-                    $files = \yii\helpers\FileHelper::findFiles(Yii::getAlias($alias . '/mail/layouts'), ['only' => ['html.php', 'text.php'], 'recursive' => true]);
+                    $files = FileHelper::findFiles(Yii::getAlias($alias . '/mail/layouts'), ['only' => ['html.php', 'text.php'], 'recursive' => true]);
                     foreach ($files as $file) {
                         $filename = str_replace('html.php', '', $file);
                         $filename = str_replace('text.php', '', $filename);
@@ -247,7 +228,7 @@ class Newsletters extends ActiveRecord
             $aliases = array_keys($extension['alias']);
             foreach ($aliases as $alias) {
                 if (is_dir(Yii::getAlias($alias . '/mail'))) {
-                    $files = \yii\helpers\FileHelper::findFiles(Yii::getAlias($alias . '/mail'), ['only' => ['*-html.php', '*-text.php'], 'recursive' => true]);
+                    $files = FileHelper::findFiles(Yii::getAlias($alias . '/mail'), ['only' => ['*-html.php', '*-text.php'], 'recursive' => true]);
                     foreach ($files as $file) {
                         $filename = str_replace('-html.php', '', $file);
                         $filename = str_replace('-text.php', '', $filename);
@@ -294,15 +275,25 @@ class Newsletters extends ActiveRecord
     /**
      * @return object of \yii\db\ActiveQuery
      */
-    public function getSubscribersFromList($cond = null)
+    public function getSubscribersFromList($cond = null, $asArray = true)
     {
         if (class_exists('\wdmg\subscribers\models\Subscribers')) {
             if ($cond) {
+
                 $list = new \wdmg\subscribers\models\Subscribers();
-                return $list::find()->select(['id', 'email'])->where($cond)->asArray()->indexBy('id')->all();
+                if ($asArray)
+                    return $list::find()->select(['id', 'email', 'name'])->where($cond)->asArray()->indexBy('id')->all();
+                else
+                    return $list::find()->where($cond)->indexBy('id')->all();
+
             } else {
+
                 $list = new \wdmg\subscribers\models\Subscribers();
-                return $list::find()->select(['id', 'email'])->asArray()->indexBy('id')->all();
+                if ($asArray)
+                    return $list::find()->select(['id', 'email', 'name'])->asArray()->indexBy('id')->all();
+                else
+                    return $list::find()->indexBy('id')->all();
+
             }
         } else {
             return null;
@@ -310,11 +301,11 @@ class Newsletters extends ActiveRecord
     }
 
     /**
-     * @return object of \yii\db\ActiveQuery
+     * @return array of recipients
      */
     public function getRecipients()
     {
-        $emails = [];
+        $recipients = [];
         $validator = new EmailValidator();
         $data = \yii\helpers\Json::decode($this->recipients);
         $validator->allowName = true;
@@ -323,14 +314,43 @@ class Newsletters extends ActiveRecord
                 preg_match("/[\._a-zA-Z0-9-]+@[\._a-zA-Z0-9-]+/i", $item, $matches);
                 if ($email = $matches[0]) {
                     if ($validator->validate($email)) {
-                        $emails[] = $email;
+
+                        $isExist = false;
+                        foreach ($recipients as $recipient) {
+                            if ($recipient['email'] == $email)
+                                $isExist = true;
+                        }
+
+                        if (!$isExist)
+                            $recipients[] = ['email' => $email];
+                        else
+                            continue;
+
                     }
                 }
             } else if (preg_match('/list_id:(\d)/', $key, $match)) {
-                if ($list = $this->getSubscribersFromList(['list_id' => intval($match[1])])) {
+                if ($list = $this->getSubscribersFromList(['list_id' => intval($match[1]), 'status' => \wdmg\subscribers\models\Subscribers::SUBSCRIBERS_STATUS_ACTIVE], false)) {
                     foreach ($list as $key => $item) {
-                        if ($validator->validate($item['email'])) {
-                            $emails[] = $item['email'];
+                        if ($validator->validate($item->email)) {
+
+                            $isExist = false;
+                            foreach ($recipients as $recipient) {
+                                if ($recipient['email'] == $item->email)
+                                    $isExist = true;
+                            }
+
+                            if (!$isExist) {
+                                $recipients[] = [
+                                    'email' => $item->email,
+                                    'name' => (isset($item->name)) ? $item->name : null,
+                                    'user_id' => (isset($item->user_id)) ? $item->user_id : null,
+                                    'username' => (isset($item->user)) ? $item->user->username : null,
+                                    'unsubscribe_url' => (isset($item->unsubscribe_link)) ? $item->unsubscribe_link : null,
+                                    'manage_url' => (isset($item->manage_link)) ? $item->manage_link : null
+                                ];
+                            } else {
+                                continue;
+                            }
                         }
                     }
                 }
@@ -338,13 +358,24 @@ class Newsletters extends ActiveRecord
                 preg_match("/[\._a-zA-Z0-9-]+@[\._a-zA-Z0-9-]+/i", $item, $matches);
                 if ($email = $matches[0]) {
                     if ($validator->validate($email)) {
-                        $emails[] = $email;
+
+                        $isExist = false;
+                        foreach ($recipients as $recipient) {
+                            if ($recipient['email'] == $email)
+                                $isExist = true;
+                        }
+
+                        if (!$isExist)
+                            $recipients[] = ['email' => $email];
+                        else
+                            continue;
+
                     }
                 }
             }
         }
 
-        return $emails;
+        return $recipients;
     }
 
     /**
@@ -394,6 +425,91 @@ class Newsletters extends ActiveRecord
         } else {
             return null;
         }
+    }
+
+    /**
+     * @return array or null
+     */
+    public function getTemplateLayouts()
+    {
+        $layouts = null;
+        if (!is_null($this->layouts)) {
+            if (is_dir(FileHelper::normalizePath(Yii::getAlias($this->layouts)))) {
+
+                if (file_exists(FileHelper::normalizePath(Yii::getAlias($this->layouts.'/html.php'))))
+                    $layouts['html'] = FileHelper::normalizePath($this->layouts . '/html');
+
+                if (file_exists(FileHelper::normalizePath(Yii::getAlias($this->layouts.'/text.php'))))
+                    $layouts['text'] = FileHelper::normalizePath($this->layouts . '/text');
+            }
+        }
+
+        return $layouts;
+    }
+
+    /**
+     * @return array or null
+     */
+    public function getTemplateViews()
+    {
+        $views = null;
+        if (!is_null(FileHelper::normalizePath(Yii::getAlias($this->views)))) {
+
+            if (file_exists(FileHelper::normalizePath(Yii::getAlias($this->views.'-html.php'))))
+                $views['html'] = FileHelper::normalizePath($this->views.'-html');
+            elseif (file_exists(FileHelper::normalizePath(Yii::getAlias($this->views.'/html.php'))))
+                $views['html'] = FileHelper::normalizePath($this->views.'/html');
+
+            if (file_exists(FileHelper::normalizePath(Yii::getAlias($this->views.'-text.php'))))
+                $views['text'] = FileHelper::normalizePath($this->views.'-text');
+            elseif (file_exists(FileHelper::normalizePath(Yii::getAlias($this->views.'/text.php'))))
+                $views['text'] = FileHelper::normalizePath($this->views.'/text');
+
+        }
+
+        return $views;
+    }
+
+
+    public function getWorkflow()
+    {
+        if (is_string($this->workflow) || !is_array($this->workflow)) {
+            if (!($this->workflow = Json::decode($this->workflow)))
+                $this->workflow = [];
+        }
+        return $this->workflow;
+    }
+
+    public function updateWorkflow($key, $param)
+    {
+        $workflow = $this->getWorkflow();
+        if (is_array($workflow)) {
+            if (is_array($param)) {
+                foreach ($param as $name => $value) {
+                    $workflow[$key][$name] = $value;
+                }
+            } else {
+                $workflow[$key] = $param;
+            }
+            $this->updateAttributes(['workflow' => Json::encode($workflow)]);
+        }
+    }
+
+
+    /**
+     * @return
+     */
+    public function getProgress()
+    {
+        if (is_array($this->workflow)) {
+
+            if (isset($this->workflow['recipients']) && isset($this->workflow['count'])) {
+                if ((count($this->workflow['recipients']) >= 0) && intval($this->workflow['count']) > 0)
+                    return (count($this->workflow['recipients']) / intval($this->workflow['count']) * 100);
+            }
+
+        }
+        return 0;
     }
 
     /**