+++ views/list/index.php
@@ -7,6 +7,7 @@ use wdmg\widgets\SelectInput;
 use wdmg\helpers\StringHelper;
 use yii\helpers\Url;
 use yii\bootstrap\Modal;
+use yii\bootstrap\Progress;
 
 /* @var $this yii\web\View */
 /* @var $searchModel wdmg\newsletters\models\NewslettersSearch */
@@ -21,10 +22,13 @@ $this->params['breadcrumbs'][] = $this->title;
 </div>
 <div class="newsletters-index">
 
-    <?php Pjax::begin(); ?>
+    <?php Pjax::begin([
+        'id' => 'newslettersAjax'
+    ]); ?>
     <?php /*echo $this->render('_search', ['model' => $searchModel]);*/ ?>
 
     <?= GridView::widget([
+        'id' => 'newslettersList',
         'dataProvider' => $dataProvider,
         'filterModel' => $searchModel,
         'layout' => '{summary}<br\/>{items}<br\/>{summary}<br\/><div class="text-center">{pager}</div>',
@@ -60,7 +64,7 @@ $this->params['breadcrumbs'][] = $this->title;
                         if (preg_match('/email_id:(\d)/', $key)) {
                             $emails[] = '<span class="label label-info">' . $item . '</span>';
                         } elseif (preg_match('/list_id:(\d)/', $key, $match)) {
-                            $count = $searchModel->getSubscribersCount(['list_id' => intval($match[1])]);
+                            $count = $searchModel->getSubscribersCount(['list_id' => intval($match[1]), 'status' => \wdmg\subscribers\models\Subscribers::SUBSCRIBERS_STATUS_ACTIVE]);
                             if (!is_null($count))
                                 $lists[] = '<span class="label label-warning">' . $item . ' ('.$count.')</span>';
                             else
@@ -116,19 +120,60 @@ $this->params['breadcrumbs'][] = $this->title;
                     'class' => 'text-center'
                 ],
                 'value' => function($data) {
-                    if ($data->status == $data::NEWSLETTERS_STATUS_ACTIVE)
-                        return '<span class="label label-success">'.Yii::t('app/modules/newsletters','Active').'</span>';
-                    elseif ($data->status == $data::NEWSLETTERS_STATUS_DISABLED)
-                        return '<span class="label label-default">'.Yii::t('app/modules/newsletters','Disabled').'</span>';
-                    else
+                    if ($data->status == $data::NEWSLETTERS_STATUS_ACTIVE) {
+                        if (isset($data->workflow['status'])) {
+                            if ($data->workflow['status'] == "run") {
+                                return Progress::widget([
+                                    'percent' => $data->progress,
+                                    'label' => round($data->progress) . '%',
+                                    'barOptions' => [
+                                        'class' => 'progress-bar-success'
+                                    ],
+                                    'options' => [
+                                        'id' => 'progress-'.$data->id,
+                                        'data-key' => $data->id,
+                                        'class' => 'active progress-striped'
+                                    ],
+                                ]);
+                            } else if ($data->workflow['status'] == "pause") {
+                                return Progress::widget([
+                                    'percent' => $data->progress,
+                                    'label' => round($data->progress) . '%',
+                                    'barOptions' => [
+                                        'class' => 'progress-bar-warning'
+                                    ],
+                                    'options' => [
+                                        'id' => 'progress-'.$data->id,
+                                        'data-key' => $data->id,
+                                        'class' => 'active progress-striped'
+                                    ],
+                                ]);
+                            } else if ($data->workflow['status'] == "stop") {
+                                return Progress::widget([
+                                    'percent' => $data->progress,
+                                    'label' => round($data->progress) . '%',
+                                    'barOptions' => [
+                                        'class' => 'progress-bar-danger'
+                                    ],
+                                    'options' => [
+                                        'id' => 'progress-'.$data->id,
+                                        'data-key' => $data->id
+                                    ],
+                                ]);
+                            }
+                        }
+
+                        return '<span class="label label-success">' . Yii::t('app/modules/newsletters', 'Active') . '</span>';
+
+                    } elseif ($data->status == $data::NEWSLETTERS_STATUS_DISABLED) {
+                        return '<span class="label label-default">' . Yii::t('app/modules/newsletters', 'Disabled') . '</span>';
+                    } else {
                         return $data->status;
+                    }
                 }
             ],
-            /*'workflow',
-            'params',*/
-            'created_at:datetime',
-            'updated_at:datetime',
-
+            /*'created_at:datetime',
+            'updated_at:datetime',*/
             [
                 'class' => 'yii\grid\ActionColumn',
                 'header' => Yii::t('app/modules/newsletters', 'Actions'),
@@ -138,35 +183,46 @@ $this->params['breadcrumbs'][] = $this->title;
                 ],
                 'buttons'=> [
                     'refresh' => function($url, $data, $key) {
-                        return Html::a('Refresh&nbsp;<span class="glyphicon glyphicon-refresh"></span>', Url::to(['process/refresh', 'id' => $data['id']]), [
+                        return Html::a(Yii::t('app/modules/newsletters', 'Refresh') . '&nbsp;<span class="glyphicon glyphicon-refresh"></span>', Url::toRoute(['process/refresh', 'id' => $data['id']]), [
                             'title' => Yii::t('app/modules/newsletters', 'Refresh progress'),
                             'data-toggle' => 'refresh-progress',
                             'data-id' => $key,
-                            'data-pjax' => '0'
-                        ]);
+                            'data-pjax' => '1'
+                        ]) . '<br/>';
                     },
                     'stop' => function($url, $data, $key) {
-                        return Html::a('Stop&nbsp;<span class="glyphicon glyphicon-stop"></span>', Url::to(['process/stop', 'id' => $data['id']]), [
+                        return Html::a(Yii::t('app/modules/newsletters', 'Stop') . '&nbsp;<span class="glyphicon glyphicon-stop"></span>', Url::toRoute(['process/stop', 'id' => $data['id']]), [
                             'title' => Yii::t('app/modules/newsletters', 'Stop newsletter'),
                             'class' => 'text-danger',
                             'data-id' => $key,
-                            'data-pjax' => '0'
-                        ]);
+                            'data-pjax' => '1'
+                        ]) . '<br/>';
                     },
                     'pause' => function($url, $data, $key) {
-                        return Html::a('Suspend&nbsp;<span class="glyphicon glyphicon-pause"></span>', Url::to(['process/pause', 'id' => $data['id']]), [
+                        return Html::a(Yii::t('app/modules/newsletters', 'Suspend') . '&nbsp;<span class="glyphicon glyphicon-pause"></span>', Url::toRoute(['process/pause', 'id' => $data['id']]), [
                             'title' => Yii::t('app/modules/newsletters', 'Suspend newsletter'),
                             'class' => 'text-warning',
                             'data-id' => $key,
-                            'data-pjax' => '0'
-                        ]);
+                            'data-pjax' => '1'
+                        ]) . '<br/>';
                     },
                     'play' => function($url, $data, $key) {
-                        return Html::a('Start&nbsp;<span class="glyphicon glyphicon-play"></span>', Url::to(['process/run', 'id' => $data['id']]), [
+                        return Html::a(Yii::t('app/modules/newsletters', 'Start') . '&nbsp;<span class="glyphicon glyphicon-play"></span>', Url::toRoute(['process/run', 'id' => $data['id']]), [
                             'title' => Yii::t('app/modules/newsletters', 'Start newsletter'),
                             'class' => 'text-success',
                             'data-id' => $key,
-                            'data-pjax' => '0'
+                            'data-pjax' => '1'
+                        ]) . '<br/>';
+                    },
+                    'preview' => function($url, $data, $key) {
+                        $url = Url::toRoute(['list/preview', 'id' => $data['id']]);
+                        return Html::a(Yii::t('app/modules/newsletters', 'Preview') . '&nbsp;<span class="glyphicon glyphicon-envelope"></span>', $url, [
+                            'class' => 'email-preview-link',
+                            'title' => Yii::t('app/modules/newsletters', 'Preview email'),
+                            'data-toggle' => 'modal',
+                            'data-target' => '#emailPreview',
+                            'data-id' => $key,
+                            'data-pjax' => '1'
                         ]);
                     },
                 ],
@@ -182,7 +238,30 @@ $this->params['breadcrumbs'][] = $this->title;
                         return \yii\helpers\Url::toRoute(['list/delete', 'id' => $key]);
 
                 },
-                'template' => '{view} {update} {delete} {refresh} {play} {stop} {pause}'
+                'visibleButtons' => [
+                    'stop' => function ($model) {
+                        if (isset($model->workflow['status'])) {
+                            return ($model->workflow['status'] == "run");
+                        }
+
+                        return false;
+                    },
+                    'pause' => function ($model) {
+                        if (isset($model->workflow['status'])) {
+                            return ($model->workflow['status'] == "run");
+                        }
+
+                        return false;
+                    },
+                    'play' => function ($model) {
+                        if (isset($model->workflow['status'])) {
+                            return !($model->workflow['status'] == "run");
+                        }
+
+                        return true;
+                    }
+                ],
+                'template' => '{view} {update} {delete} <br/> {refresh} {play} {stop} {pause} {preview}'
             ],
         ],
         'pager' => [
@@ -207,5 +286,29 @@ $this->params['breadcrumbs'][] = $this->title;
     </div>
     <?php Pjax::end(); ?>
 </div>
+
+<?php $this->registerJs(<<< JS
+    $('body').delegate('.email-preview-link', 'click', function(event) {
+        event.preventDefault();
+        $.get(
+            $(this).attr('href'),
+            function (data) {
+                $('#emailPreview .modal-body').html(data);
+                $('#emailPreview').modal();
+            }  
+        );
+    });
+JS
+); ?>
+
+<?php Modal::begin([
+    'id' => 'emailPreview',
+    'size' => 'modal-lg',
+    'header' => '<h4 class="modal-title">'.Yii::t('app/modules/newsletters', 'Email preview').'</h4>',
+    'clientOptions' => [
+        'show' => false
+    ]
+]); ?>
+<?php Modal::end(); ?>
 
 <?php echo $this->render('../_debug'); ?>